FSCDiff-main/requirements.txt
accelerate==0.17.1
albumentations==1.3.0
einops==0.6.0
matplotlib==3.6.2
numba==0.56.4
omegaconf==2.2.3
opencv-python==4.6.0.66
Pillow==9.2.0
# Editable install with no version control (pix2tex==0.0.29)
scikit-learn==1.1.3
scipy==1.9.3
timm==0.6.12

wandb==0.13.10

numpy==1.23.4
tqdm==4.64.1
argparse==1.4.0
setuptools==65.5.0
ema_pytorch
pytorch_fid
mmcv-full